Removed require 'dotenv' since it's not used (?).
I had a problem with discourse_api gem on a new Rails app -
my app just crashed with this error:
```
/Users/user/.rvm/gems/ruby-2.2.2/gems/discourse_api-0.4.3/lib/discourse_api/client.rb:1:in `require': cannot load such file -- dotenv (LoadError)
```

I think it's because dotenv is a development dependency in gemspec,
yet we have a require 'dotenv' in client.rb. It's never loaded by
bundler when the gem is just added to a gem file, so the app just crashes.

I don't see it being used in client.rb, so I just deleted it.
It's used in examples/groups.rb so I've added the require there.

I'm not sure if we need dotenv at all, can someone comment on it?