Ensure tests don't require discourse to be running
Eventually I would like to create some integration tests that actually talk
to a discourse instance, but for these current rspec tests we need them
to run without requiring a local instance of discourse to be running and
to be in just the right state for the tests to pass. Not having the
tests actually talk to discourse will also let us us Travis for testing.

The following changes were made:

- stub out the create category test
- broke up some really long lines